Removed WinHttpHandler from System.Net.Http.
