+++ src/libraries/Common/src/System/Net/Http/WinInetProxyHelper.cs
